Polish the change.
