+++ sha1.cpp
@@ -2,22 +2,4 @@
     sha1.cpp - This file is intenionally left blank.
     This file is no longer needed. It is only kept for backard compatibility.
     Simply: #include "sha1.hpp"
-
-    ============
-    SHA-1 in C++
-    ============
-
-    100% Public Domain.
-
-    Original C Code
-        -- Steve Reid <steve@edmweb.com>
-    Small changes to fit into bglibs
-        -- Bruce Guenter <bruce@untroubled.org>
-    Translation to simpler C++ Code
-        -- Volker Diels-Grabsch <v@njh.eu>
-    Safety fixes
-        -- Eugene Hopkinson <slowriot at voxelstorm dot com>
-    Header-only library
-        -- Zlatko Michailov <zlatko@michailov.org>
 */
-
